Event controller

+++ server/src/main/java/server/database/EventRepository.java
@@ -19,4 +19,4 @@ import commons.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EventRepository extends JpaRepository<Event, Long> {}
+public interface EventRepository extends JpaRepository<Event, Integer> {}
